chore(1-3336): move traffic calculations to utils
